Fix the bug for extracting path to be appended to sys.path
Previously I used rstrip('sadit'). The intention was to remove the
saidt suffix, but the implementation was to remove all chars of sadit.
It will cause problem if the path happens to contain these characters.

It also update the url of networkx in install script. The old url has
become invalid.

TESTED=./cmdsadit sim -c ./Example/SimExample.py